refactor: moved code to proper files
