+++ src/files/utils.rs
@@ -1,5 +1,6 @@
 use crate::constants::FILENAME_SEPARATORS;
 use crate::constants::SUPPORTED_EXTENSIONS;
+use crate::utils::string::levenshtein_distance;
 use std::{
     ffi::{OsStr, OsString},
     path::Path,
@@ -59,6 +60,59 @@ pub fn get_stem_name_trimmed_at_first_separator(file_name: &OsStr) -> OsString {
         .trim_end_matches(|c: char| c.is_ascii_digit())
         .to_string();
     OsString::from(x)
+}
+
+/// Find similar repo names in the data directory based on fuzzy matching
+pub fn find_similar_repos(data_dir: &Path, target_repo: &str) -> String {
+    let mut similar_repos = Vec::new();
+
+    if let Ok(entries) = std::fs::read_dir(data_dir) {
+        for entry in entries.flatten() {
+            if let Ok(file_type) = entry.file_type() {
+                if file_type.is_dir() {
+                    if let Some(username) = entry.file_name().to_str() {
+                        // Check subdirectories (repos) within each user directory
+                        let user_dir = data_dir.join(username);
+                        if let Ok(repo_entries) = std::fs::read_dir(user_dir) {
+                            for repo_entry in repo_entries.flatten() {
+                                if let Ok(repo_file_type) = repo_entry.file_type() {
+                                    if repo_file_type.is_dir() {
+                                        if let Some(repo_name) = repo_entry.file_name().to_str() {
+                                            let full_repo = format!("{}/{}", username, repo_name);
+
+                                            // Calculate similarity
+                                            let distance =
+                                                levenshtein_distance(target_repo, &full_repo);
+                                            let max_len =
+                                                std::cmp::max(target_repo.len(), full_repo.len());
+
+                                            // Consider repos with distance <= 3 or similarity >= 70%
+                                            if distance <= 3
+                                                || (max_len > 0
+                                                    && distance as f32 / max_len as f32 <= 0.3)
+                                            {
+                                                similar_repos.push(full_repo);
+                                            }
+                                        }
+                                    }
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+        }
+    }
+
+    // Sort by similarity (lower distance = more similar)
+    similar_repos.sort_by(|a, b| {
+        let dist_a = levenshtein_distance(target_repo, a);
+        let dist_b = levenshtein_distance(target_repo, b);
+        dist_a.cmp(&dist_b)
+    });
+
+    // Return only the top entry as a string
+    similar_repos.into_iter().next().unwrap_or_default()
 }
 
 #[cfg(test)]
